+++ src/main/java/git/dragomordor/simpletms/fabric/SimpleTMsMod.java
@@ -4,6 +4,8 @@ import com.google.gson.JsonObject;
 import git.dragomordor.simpletms.fabric.config.SimpleTMsConfig;
 import git.dragomordor.simpletms.fabric.item.SimpleTMsItemGroups;
 import git.dragomordor.simpletms.fabric.item.SimpleTMsItems;
+import git.dragomordor.simpletms.fabric.util.LootTables;
+import git.dragomordor.simpletms.fabric.util.TMsTRsList;
 import net.fabricmc.api.ModInitializer;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -11,10 +13,6 @@ import org.apache.logging.log4j.Logger;
 public class SimpleTMsMod implements ModInitializer {
     public static final Logger LOGGER = LogManager.getLogger(SimpleTMsMod.class); // create logger
     public static final String MODID = "simpletms"; // mod ID
-
-    // Recipes
-    public static JsonObject TR_RECIPE = null;
-    public static JsonObject TM_RECIPE = null;
 
     @Override
     public void onInitialize() {
@@ -25,6 +23,10 @@ public class SimpleTMsMod implements ModInitializer {
         // Register creative tabs
         SimpleTMsItemGroups.registerItemGroups(); //register TM item tab
 
+        // Register TM list
+        TMsTRsList.registerTMList();
+        // Register events
+        LootTables.registerEvents();
     }
 }