+++ changelog.md
@@ -1,2 +1,10 @@
 ## [1.0.0 (January 7th, 2024)](#1-0-0)
 - Initial Release
+
+#### [1.0.1 (January 8th, 2024)](#1-0-1)
+- Fixed Compatability issues between version  1.4.1 of Cobblemon
+
+#### [1.1.0 (January 16th, 2024)](#1-1-0)
+- Added 2 config options for TM and TR drop chance from wild Pokemon
+- Pokémon now have a (by default) 5% chance to drop a TR or a 0.01% chance to drop a TM.
+TMs and TRs that drop are based on the Pokémon's tm, egg and tutor move list.